@@ -4,10 +4,6 @@
 require_relative 'logger_formatter'
 
 if AppOpticsAPM.loaded && defined?(Lumberjack::Formatter)
-  module Lumberjack
-    class Formatter
-      prepend AppOpticsAPM::Logger::Formatter
-    end
-  end
+  Lumberjack::Formatter.send(:prepend, AppOpticsAPM::Logger::Formatter)
 end
 
