@@ -16,6 +16,7 @@ end if ENV["SIMPLECOV_COVERAGE"]
 
 require 'rubygems'
 require 'bundler/setup'
+require 'fileutils'
 require 'minitest/spec'
 require 'minitest/autorun'
 require 'minitest/reporters'
@@ -110,17 +111,10 @@ end
 case File.basename(ENV['BUNDLE_GEMFILE'])
 when /delayed_job/
   require './test/servers/delayed_job'
-
-when /rails[56]/
+when /rails/
   require './test/servers/rails5x_8140'
-  sleep 1
-  require './test/servers/rails5x_api_8150'
-when /rails4/
-  require './test/servers/rails4x_8140'
-
 when /frameworks/
 when /libraries/
-
   # Load Sidekiq for libaries tests
   # use `export NO_SIDEKIQ=true` to stop sidekiq from loading
   # when running individual test files
@@ -171,6 +165,28 @@ def get_all_traces
     AppOpticsAPM::Reporter.get_all_traces
   else
     []
+  end
+end
+
+##
+# read the ActiveRecord logfile and match it with regex
+# use case: test if trace-id has been injected in query
+#
+# `clear_query_log` before next test
+def query_logged?(regex)
+  File.open(ENV['QUERY_LOG_FILE']).read() =~ regex
+end
+
+##
+# clear the ActiveRecord logfile, but don't remove it
+# create if it doesn't exist
+#
+def clear_query_log
+  ENV['QUERY_LOG_FILE'] ||= '/tmp/query_log.txt'
+  if File.exist?(ENV['QUERY_LOG_FILE'])
+    File.truncate(ENV['QUERY_LOG_FILE'], 0)
+  else
+    FileUtils.touch(ENV['QUERY_LOG_FILE'])
   end
 end
 
